Fix编译错误和警告：修复 MegolmService 方法缺失，清理未使用导入和变量，并优化 E2EE 路由

+++ src/storage/voice.rs
@@ -54,7 +54,7 @@ impl VoiceUsageStatsStorage {
         user_id: &str,
         days: i64,
     ) -> Result<Vec<VoiceUsageStats>, sqlx::Error> {
-        let date_threshold = chrono::Utc::now() - chrono::Duration::days(days as i64);
+        let date_threshold = chrono::Utc::now() - chrono::Duration::days(days);
 
         let rows = sqlx::query(
             r#"
